# backend/api/src/main/java/dev/byowa/hdp/dto/AuthResponse.java
package dev.byowa.hdp.dto;

public class AuthResponse {
    private String token;

    public AuthResponse() {}

    public AuthResponse(String token) {
        this.token = token;
    }

    public String getToken() {
        return token;
    }

    public void setToken(String token) {
        this.token = token;
    }
}